+++ src/main/java/org/jks/persistence/ArticleDao.java
@@ -6,13 +6,13 @@ import java.util.List;
 import org.jks.domain.Article;
 import org.jks.persistence.common.GenericDao;
 
-public interface ArticleDao  extends GenericDao<Article, Integer>  {
+public interface ArticleDao  extends GenericDao<Article, Long>  {
 
 
 	public Article getArticleByTimestamp(Timestamp datearticle);
 	
 	public Article getArticleBySubject(String subject);
 	
-	public List<Article> getArticlesInSection(Integer section);
+	public List<Article> getArticlesInSection(long section);
 
 }
