+++ src/main/java/org/jks/persistence/SectionDao.java
@@ -3,7 +3,7 @@ import org.jks.domain.Section;
 import org.jks.domain.User;
 import org.jks.persistence.common.GenericDao;
 
-public interface SectionDao extends GenericDao<Section, Integer>  {
+public interface SectionDao extends GenericDao<Section, Long>  {
 
 	public Section getSectionByName(String sectionArticle);
